refactor(domain/dateTime): 상영 날짜 및 시간에 대한 리스트를 반환해주는 클래스의 이름 변경
