+++ domain/src/main/java/com/example/domain/dateTime/RunningDate.kt
@@ -1,9 +1,9 @@
-package com.example.domain.setter
+package com.example.domain.dateTime
 
 import java.time.LocalDate
 import java.time.temporal.ChronoUnit.DAYS
 
-class RunningDateSetter {
+class RunningDate {
     fun getRunningDates(start: LocalDate, end: LocalDate): List<LocalDate> {
         val betweenDays = DAYS.between(start, end)
         return (ZERO..betweenDays).map { start.plusDays(it) }
